requirements.txt
Pillow==10.0.1
python_Levenshtein==0.21.1
Requests==2.31.0
telepot==12.7
watchdog==3.0.0